+++ sponsoredTransactions/frontend/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased changes
 
+## 2.1.0
+
+- Migrate dApp from using deprecated JSON-RPC client to new gRPC client.
+
 ## 2.0.0
 
 - Remove tab to register public key